+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.7.2 Release
 * Fix NPE for armor stand hologram location is null
 * Fixed issue when sign instances section not found in file
+* Fixed arena join permission
 
 ### 1.7.1 Release (04.01.2021)
 * Changed summary message for one murderer round