+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed issue when sign instances section not found in file
 * Fixed arena join permission
 * Added 1.11+ compatibility
+* Added murderer and detective shuffler to avoid the same characters when you are only playing with the same people
 
 ### 1.7.1 Release (04.01.2021)
 * Changed summary message for one murderer round