+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-### 1.7.2 Release
+### 1.7.2 Release (10.01.2021 - 18.01.2021)
 * Fix NPE for armor stand hologram location is null
 * Fixed issue when sign instances section not found in file
 * Fixed arena join permission
 * Added 1.11+ compatibility
 * Added murderer and detective shuffler to avoid the same characters when you are only playing with the same people
+* Added option to disable death message completely 
 
 ### 1.7.1 Release (04.01.2021)
 * Changed summary message for one murderer round